Make first commit of the BigMath library, with basic trig functions
This library is intended for those wanting to preform calculations
at a customizable level of precision. This library will work alongside
the basic arbitrary precision classes provided by Java; BigDecimal,
MathContext, and possibly BigInteger in the future.

Currently, the library includes the trignometric functions sine, cosine,
and tangent, as well as factorial (using a very ineffecient method it
may be added), and degrees to radians, and radians to degrees conversion
functions.

The hope of this library is to amass a greater deal of functions and
refactoring to offer a simple, but elegant solution to those wishing
to have more control over their use of mathematics