Begin implementation of method for calculating pi to arbitrary precision
This method will be an infinite series, allowing the programmer to decide
how many iterations they'd like to allow to get as many digits of pi as he
or she would like.

So far, the method gets approaches zero, not pi. The page this equation
was pulled of was here:
http://math.stackexchange.com/questions/716944/how-to-find-continued-fraction-of-pi